backend/admin-service/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
python-decouple==3.8
gunicorn==22.0.0
